requirements.txt
flask==2.0.1
flask-cors==3.0.10
bcrypt==3.2.0
mysql-connector-python==8.0.26
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
Flask-MySQLdb==2.0.0
Flask-WTF==1.2.2
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.1.0
Werkzeug==3.1.3
WTForms==3.2.1
zipp==3.21.0
